add entities
pretty much everything that can be interacted with in the game is an entity, defined by this struct. any field we have here will be in all entities. This is an ok compromise for now, but as the game becomes more complex we may instead want to look at a full blown entity-component-system pattern for efficiency.